feat(k8s/volumes): move storages table to volumes view
