+++ app/kubernetes/views/volumes/components/applications-storages-datatable/template.html
