+++ app/kubernetes/views/volumes/volumes.js
@@ -2,4 +2,7 @@ angular.module('portainer.kubernetes').component('kubernetesVolumesView', {
   templateUrl: './volumes.html',
   controller: 'KubernetesVolumesController',
   controllerAs: 'ctrl',
+  bindings: {
+    $transition$: '<',
+  },
 });
